Fix: Change file size type from Int to Long in picture update errors
The `FileSizeOutOfBounds` error class and its related implementations used `Int` to represent file sizes. This could lead to integer overflow issues with larger files.

This commit changes the `size`, `min`, and `max` properties to `Long` in the following error classes to correctly handle larger file sizes:

- `UpdatePictureRemoteDataSourceError.FileSizeOutOfBounds`
- `UpdatePictureRepositoryError.FileSizeOutOfBounds`
- `UpdatePictureError.FileSizeOutOfBounds`

+++ app/src/main/java/timur/gilfanov/messenger/data/source/remote/UpdatePictureRemoteDataSourceError.kt
@@ -1,7 +1,7 @@
 package timur.gilfanov.messenger.data.source.remote
 
 sealed interface UpdatePictureRemoteDataSourceError : RemoteUserDataSourceError {
-    data class FileSizeOutOfBounds(val size: Int, val min: Int, val max: Int) :
+    data class FileSizeOutOfBounds(val size: Long, val min: Long, val max: Long) :
         UpdatePictureRemoteDataSourceError
 
     data class WidthOutOfBounds(val width: Int, val min: Int, val max: Int) :
